% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_import_tre_detail.R
\name{admin_update_tre}
\alias{admin_update_tre}
\title{Mise à jour des TRE}
\usage{
admin_update_tre(path_excel, annee, output_folder)
}
\arguments{
\item{path_excel}{chemin du fichier}

\item{annee}{est annee du dernier TRE disponible}

\item{output_folder}{est le dossier de sortie}
}
\description{
Mise à jour des TRE
}
