% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_chainage.R
\name{agregat_annuel}
\alias{agregat_annuel}
\title{Agrégat annuel répété (alias pour compatibilité)}
\usage{
agregat_annuel(x)
}
\arguments{
\item{x}{Vecteur trimestriel.}
}
\value{
Vecteur de même longueur.
}
\description{
Agrégat annuel répété (alias pour compatibilité)
}
