% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01bis_projection_ratios.R
\name{agreger_tre_etalonnage}
\alias{agreger_tre_etalonnage}
\title{Agréger le TRE vers la nomenclature d'étalonnage}
\usage{
agreger_tre_etalonnage(db_tre, table_passage)
}
\arguments{
\item{db_tre}{La base TRE brute (Niveau 2 Branches)}

\item{table_passage}{Dataframe avec colonnes : Code_Branche_Niv2, Code_Branche_Etal}
}
\description{
Agréger le TRE vers la nomenclature d'étalonnage
}
