% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_deflate.R
\name{apply_price_to_bench}
\alias{apply_price_to_bench}
\title{Appliquer Prix sur résultat de Benchmarking}
\usage{
apply_price_to_bench(df_bench, df_prix, operation = c("deflate", "inflate"))
}
\arguments{
\item{df_bench}{Données de benchmarking (doit contenir 'full_code')}

\item{df_prix}{Données de prix (Code_Produit/IP ou branche_prix/prix)}

\item{operation}{"deflate" ou "inflate"}
}
\description{
Cette fonction réalise l'opération de déflation ou d'inflation.
Elle est compatible avec les anciens formats (branche_prix/prix) et les nouveaux (Code_Produit/IP).
}
