% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_import_nomenclatures.R
\name{build_nomenclatures}
\alias{build_nomenclatures}
\title{Construire les tables de passage (nomenclatures) et les enregistrer en RDS}
\usage{
build_nomenclatures(path_excel, output_folder, overwrite = FALSE)
}
\arguments{
\item{path_excel}{Chemin complet vers le fichier Excel des nomenclatures.}

\item{output_folder}{Dossier de sortie des fichiers RDS.}

\item{overwrite}{Logique. Si FALSE et que les fichiers existent déjà, stop.}
}
\value{
Liste invisible contenant \code{Map_Branches} et \code{Map_Produits}.
}
\description{
Génère :
\itemize{
\item \code{Map_Branches.rds} à partir de l'onglet \code{Nomen_Bran}
\item \code{Map_Produits.rds} à partir de l'onglet \code{Nomen_Prod}
}

Cette fonction est destinée à être utilisée uniquement lors d'une mise à jour
exceptionnelle des nomenclatures Excel. Les fichiers RDS produits sont ensuite
réutilisés dans les compilations courantes.

Structure attendue de l'onglet \code{Nomen_Prod} :
A = Code_Prod_N3
B = Libelle_N3
C = Code_Prod_N2
D = Libelle_N2
E = Code_Prod_Etal
F = Code_Prod_Ct
G = Code_Prod_Tpub
H = Libelle_Tpub
I = Type_Prod

Structure attendue de l'onglet \code{Nomen_Bran} :
A = Code_Bran_N2
B = Libelle_N2
C = Code_Bran_Etal
D = Code_Bran_Ct
E = Code_Bran_Tpub
F = Libelle_Tpub
}
