% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_chainage.R
\name{calcul_indice_volume_annuel}
\alias{calcul_indice_volume_annuel}
\title{Calculer les indices de volume (annuel)}
\usage{
calcul_indice_volume_annuel(IndCrt, IndVol)
}
\arguments{
\item{IndCrt}{Flux annuels à prix courants.}

\item{IndVol}{Flux annuels en volume aux prix de l'année précédente.}
}
\value{
Vecteur des maillons annuels.
}
\description{
Maillons annuels pour une série de flux :
I_1 = 1 ; I_t = Vvol_t / Vcrt_(t-1).
}
