% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_chainage.R
\name{calcul_indice_volume_trim}
\alias{calcul_indice_volume_trim}
\title{Calculer les indices de volume trimestriels (flux)}
\usage{
calcul_indice_volume_trim(IndCrt, IndVol)
}
\arguments{
\item{IndCrt}{Flux trimestriels courants.}

\item{IndVol}{Flux trimestriels aux prix de l'année précédente.}
}
\value{
Maillons trimestriels i_q.
}
\description{
i_q = (Vvol_q / Vcrt_ann(année-1)) * 4
avec pour l'année de base : denom = Vcrt_ann(base).
}
