% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_chainage.R
\name{calcul_valeur_chainee_annuel}
\alias{calcul_valeur_chainee_annuel}
\title{Calculer une valeur chaînée annuelle (flux)}
\usage{
calcul_valeur_chainee_annuel(IndCrt, IndVol, base_position = 1)
}
\arguments{
\item{IndCrt}{Flux annuels à prix courants.}

\item{IndVol}{Flux annuels en volume aux prix de l'année précédente.}

\item{base_position}{Position de l'année de base dans le vecteur.}
}
\value{
Valeurs chaînées annuelles.
}
\description{
Calculer une valeur chaînée annuelle (flux)
}
