% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/12_pipeline_ci_branches.R
\name{calculer_ci_branches}
\alias{calculer_ci_branches}
\title{Calculer la CI trimestrielle par branche}
\usage{
calculer_ci_branches(prod_complete, res_trim, ci_crt, ci_vol)
}
\arguments{
\item{prod_complete}{Table de production trimestrielle complète.}

\item{res_trim}{Résultat du module TRE.}

\item{ci_crt}{CI annuelle à prix courants.}

\item{ci_vol}{CI annuelle en volume.}
}
\value{
Liste contenant les objets intermédiaires et finaux du calcul de CI.
}
\description{
Calculer la CI trimestrielle par branche
}
