% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_analyse_structurelle.R
\name{calculer_coef_technique}
\alias{calculer_coef_technique}
\title{Calculer les Coefficients Techniques par Branche}
\usage{
calculer_coef_technique(df_tre)
}
\arguments{
\item{df_tre}{Le data.frame issu de Base_TRE_Historique.rds}
}
\value{
Un data.frame avec le Coefficient Technique par Branche, Année et Type de Prix.
}
\description{
Cette fonction calcule le rapport CI / Production pour chaque branche.
Elle utilise les lignes "Total" générées lors de l'import.
}
