% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/14_pipeline_comptes_branches.R
\name{calculer_comptes_branches}
\alias{calculer_comptes_branches}
\title{Calculer les comptes trimestriels par branche}
\usage{
calculer_comptes_branches(p1_agreg, p2_final_cal)
}
\arguments{
\item{p1_agreg}{Production agrégée par branche.}

\item{p2_final_cal}{CI trimestrielle finale par branche.}
}
\value{
Table synthétique des comptes par branche : P1, P2, VA.
}
\description{
Calculer les comptes trimestriels par branche
}
