% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/15_pipeline_p1p2_produits.R
\name{calculer_p1_p2_par_produit}
\alias{calculer_p1_p2_par_produit}
\title{Calculer P1 et P2 par produit}
\usage{
calculer_p1_p2_par_produit(p1_agreg, p2_final_cal, poids_p1p2)
}
\arguments{
\item{p1_agreg}{Production agrégée par branche.}

\item{p2_final_cal}{CI finale par branche.}

\item{poids_p1p2}{Poids branche-produit trimestriels.}
}
\value{
Liste contenant \code{p1_par_produit} et \code{p2_par_produit}.
}
\description{
Calculer P1 et P2 par produit
}
