% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_analyse_structurelle.R
\name{calculer_poids_matrice}
\alias{calculer_poids_matrice}
\title{Calculer la Structure (Poids) d'une Matrice}
\usage{
calculer_poids_matrice(df_tre, op_cible = "P1")

calculer_poids_matrice(df_tre, op_cible = "P1")
}
\arguments{
\item{df_tre}{Le data.frame complet.}

\item{op_cible}{"P1" pour Production, "P2" pour CI}
}
\value{
Un data.frame détaillé avec une colonne 'Poids'.
}
\description{
Calcule la part de chaque produit dans le total de la branche.
Poids = Valeur(Branche, Produit) / Valeur(Branche, Total)
}
