% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01bis_projection_ratios.R
\name{calculer_ratios_annuels_etal}
\alias{calculer_ratios_annuels_etal}
\title{Calculer les Coefficients Techniques et Poids (Base Agrégée)}
\usage{
calculer_ratios_annuels_etal(db_tre_etal)
}
\arguments{
\item{db_tre_etal}{Base TRE agrégée au niveau étalonnage (Annee, Type_Prix, Operation, Code_Branche, Code_Produit, Valeur).}
}
\description{
Calculer les Coefficients Techniques et Poids (Base Agrégée)
}
