% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_chainage.R
\name{chainer_indices}
\alias{chainer_indices}
\title{Chaîner des indices (produit cumulatif)}
\usage{
chainer_indices(indices)
}
\arguments{
\item{indices}{Vecteur de maillons.}
}
\value{
Vecteur d'indices chaînés.
}
\description{
Chaîner des indices (produit cumulatif)
}
