% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/17_modele_equilibrage_ere.R
\name{charger_modele_equilibrage_ere}
\alias{charger_modele_equilibrage_ere}
\title{Charger le modèle de bouclage ERE depuis un fichier RDS}
\usage{
charger_modele_equilibrage_ere(
  path_rds = NULL,
  output_file = "Modele_Equilibrage_ERE.rds",
  validate = TRUE
)
}
\arguments{
\item{path_rds}{Chemin du fichier \code{.rds}. Si \code{NULL}, la fonction tente
d'utiliser \code{load_config()$root_dir/Modele_Equilibrage_ERE.rds}.}

\item{output_file}{Nom du fichier si \code{path_rds = NULL}.}

\item{validate}{Si \code{TRUE}, vérifie la structure minimale de l'objet.}
}
\value{
Une liste de paramétrage du modèle de bouclage ERE.
}
\description{
Charge un fichier \code{.rds} de paramétrage du bouclage ERE.
}
