% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_utils.R
\name{clean_res}
\alias{clean_res}
\title{Consolidation des Resultats de benchmarking}
\usage{
clean_res(df, type_traitement)
}
\arguments{
\item{df}{la base de fichers benchmarkes}

\item{type_traitement}{est le type de traitement concerne}
}
\description{
Consolidation des Resultats de benchmarking
}
