% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/17_modele_equilibrage_ere.R
\name{composantes_autorisees}
\alias{composantes_autorisees}
\title{Extraire les composantes autorisées pour un produit ERE}
\usage{
composantes_autorisees(modele_obj, code_produit, autorise_seulement = TRUE)
}
\arguments{
\item{modele_obj}{Objet retourné par \code{importer_modele_equilibrage_ere_excel()}
ou \code{charger_modele_equilibrage_ere()}.}

\item{code_produit}{Code produit ERE.}

\item{autorise_seulement}{Si \code{TRUE}, ne renvoie que les composantes autorisées.}
}
\value{
Un tibble.
}
\description{
Retourne les composantes autorisées par le modèle de bouclage d'un produit.
}
