% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_chainage.R
\name{dechainer_valeurs}
\alias{dechainer_valeurs}
\title{Déchaînage des valeurs (VPAP / PaP implicite)}
\usage{
dechainer_valeurs(CntCrt, CntVolchaine, trim = TRUE)
}
\arguments{
\item{CntCrt}{Valeurs courantes (trimestrielles si trim=TRUE).}

\item{CntVolchaine}{Valeurs chaînées (trimestrielles si trim=TRUE).}

\item{trim}{TRUE pour trimestriel, FALSE pour annuel.}
}
\value{
VPAP (même longueur).
}
\description{
Calcule les valeurs aux prix de l'année précédente (VPAP) à partir de :
\itemize{
\item valeurs courantes CntCrt
\item valeurs chaînées CntVolchaine
}

Logique trimestrielle CNT :
VPAP_q = CntVolchaine_q * PaP_(annee-1)
oou PaP_(annee-1) = (TotalAnnuelCrt_(annee-1)) / (TotalAnnuelCh_(annee-1))

=> Pour une année incomplète (ex: 2024), on utilise PaP de 2023 pour 2024Q1-Q3.
}
