% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/17_ere_equilibrage_multivarie.R
\name{equilibrer_ere_multivarie}
\alias{equilibrer_ere_multivarie}
\title{Equilibrer plusieurs produits ERE (wrapper)}
\usage{
equilibrer_ere_multivarie(
  data_ere,
  model_equil,
  forcer_coherence = TRUE,
  call_cholette = rjd3bench::multivariatecholette
)
}
\arguments{
\item{data_ere}{Table longue multi-produits.}

\item{model_equil}{Table avec colonnes \code{Code_Produit} et
\code{composante_ajustable}.}

\item{forcer_coherence}{Recaler les contraintes contemporaines.}

\item{call_cholette}{Fonction solveur injectable.}
}
\value{
Liste nommee par \code{Code_Produit}.
}
\description{
Equilibrer plusieurs produits ERE (wrapper)
}
