% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10_pipeline_tre.R
\name{executer_module_tre}
\alias{executer_module_tre}
\title{Exécuter le module TRE}
\usage{
executer_module_tre(donnees, cfg)
}
\arguments{
\item{donnees}{Liste retournée par \code{\link[=charger_donnees_cnt]{charger_donnees_cnt()}}.}

\item{cfg}{Liste de configuration.}
}
\value{
Une liste nommée contenant :
\describe{
\item{ratios_annuels}{Ratios annuels calculés sur le TRE agrégé}
\item{res_trim}{Sortie complète de \code{trimestrialiser_cnt_complet()}}
\item{coef_tech_trim}{Coefficients techniques trimestriels}
\item{poids_trim}{Poids trimestriels}
\item{ct_trim}{Coefficients techniques trimestriels bruts}
}
}
\description{
Exécuter le module TRE
}
