% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11_pipeline_production_directe.R
\name{executer_optique_production_directe}
\alias{executer_optique_production_directe}
\title{Exécuter l'optique production directe}
\usage{
executer_optique_production_directe(donnees, tre)
}
\arguments{
\item{donnees}{Liste chargée par \code{\link[=charger_donnees_cnt]{charger_donnees_cnt()}}.}

\item{tre}{Résultat de \code{\link[=executer_module_tre]{executer_module_tre()}}.}
}
\value{
Une liste contenant les sorties intermédiaires et finales :
\describe{
\item{cnt_ind1_crt, cnt_ind1_ch}{Benchmarking Ind1}
\item{cnt_ind2_crt, cnt_ind2_ch}{Benchmarking Ind2}
\item{cnt_ind3_crt, cnt_ind3_ch}{Benchmarking Ind3}
\item{est_ind2_vol, est_ind3_val}{Étapes prix}
\item{df_all_crt, df_all_ch}{Tables consolidées}
\item{prod_complete}{Production trimestrielle complète}
}
}
\description{
Exécuter l'optique production directe
}
