% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_utils.R
\name{extract_target_code}
\alias{extract_target_code}
\title{Nettoie le code complet pour trouver la cible CNA}
\usage{
extract_target_code(code)
}
\arguments{
\item{code}{le code concatene dont on veut extraire seulement le code branche
Ex: "Ind1_PRIVE\emph{A01007" -> "PRIVE}A01007"}
}
\description{
Nettoie le code complet pour trouver la cible CNA
}
