% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/16_impots_taxes.R
\name{extraire_ind_impots}
\alias{extraire_ind_impots}
\title{Extraire l'indicateur Impots_Nets depuis ind_crt (IndCrt)}
\usage{
extraire_ind_impots(ind_crt, ind_cst)
}
\arguments{
\item{ind_crt}{Tibble issu de \code{import_matrix_cnt()} sur IndCrt.}

\item{ind_cst}{Tibble issu de \code{import_matrix_cnt()} sur IndCst.}
}
\value{
Liste avec :
\describe{
\item{ind_imp_crt}{Indicateur courant Impots_Nets (IndCrt).}
\item{ind_taxes_crt}{Indicateur courant Taxes (IndCrt).}
\item{ind_subv_crt}{Indicateur courant Subvention (IndCrt).}
}
}
\description{
Filtre les lignes de \code{ind_crt} dont \code{type_ind} vaut
\code{"Impots_Nets"}, \code{"Taxes"} ou \code{"Subvention"}
(tels qu'ils apparaissent en ligne 2 des en-tetes Excel).
Retourne un tibble au format attendu par \code{benchmark_groupe()}.
}
