% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_get_data.R
\name{get_tre_matrix}
\alias{get_tre_matrix}
\title{Récupérer un TRE spécifique pour calculs}
\usage{
get_tre_matrix(
  annee_cible,
  type_prix_cible = "Courant",
  matrice = "Production"
)
}
\arguments{
\item{annee_cible}{Année voulue}

\item{type_prix_cible}{"Courant" ou "Constant"}

\item{matrice}{"Production", "CI", "Demande_Finale", "VAB"}
}
\description{
Récupérer un TRE spécifique pour calculs
}
