% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_import.R
\name{import_cna_ere_structured}
\alias{import_cna_ere_structured}
\title{Importation structurée de l'Équilibre Ressources-Emplois (CnaEre)}
\usage{
import_cna_ere_structured(path, sheet = "CnaEre")
}
\arguments{
\item{path}{Chemin du fichier Excel}

\item{sheet}{Nom de la feuille (Défaut "CnaEre")}
}
\description{
Importation structurée de l'Équilibre Ressources-Emplois (CnaEre)
}
