% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_import.R
\name{import_matrix_cnt}
\alias{import_matrix_cnt}
\title{Importation robuste des matrices CNT (Format 3 lignes d'en-tête)}
\usage{
import_matrix_cnt(path, sheet)
}
\arguments{
\item{path}{chemin}

\item{sheet}{la feuille concernée}
}
\description{
Importation robuste des matrices CNT (Format 3 lignes d'en-tête)
}
