% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_import.R
\name{import_matrix_prix}
\alias{import_matrix_prix}
\title{Import spécifique pour les Prix}
\usage{
import_matrix_prix(path, sheet)
}
\arguments{
\item{path}{est le chemin du fichier}

\item{sheet}{est la feuille concernée}
}
\description{
Import spécifique pour les Prix
}
