% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_import.R
\name{import_matrix_tre}
\alias{import_matrix_tre}
\title{Importation des matrices de ratios techniques (TRE)}
\usage{
import_matrix_tre(path, sheet)
}
\arguments{
\item{path}{est le chemin du fichier}

\item{sheet}{est la feuille concernée}
}
\description{
Importation des matrices de ratios techniques (TRE)
}
