% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_import.R
\name{import_prix_niv3}
\alias{import_prix_niv3}
\title{Importation des Prix détaillés (Niveau 3)}
\usage{
import_prix_niv3(path)
}
\arguments{
\item{path}{est le chemin du fichier}
}
\description{
Importation des Prix détaillés (Niveau 3)
}
