% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_outputs.R
\name{plot_benchmark_compare}
\alias{plot_benchmark_compare}
\title{Graphique de comparaison Indicateur vs Résultat Calé}
\usage{
plot_benchmark_compare(df_source, df_bench, code_branche)
}
\arguments{
\item{df_source}{les indicateurs sources utilisés}

\item{df_bench}{les agregats calé}

\item{code_branche}{est la brahnce dont on faire l'etalonnage graphique}
}
\description{
Graphique de comparaison Indicateur vs Résultat Calé
}
