% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_import.R
\name{prepare_price_data}
\alias{prepare_price_data}
\title{Préparation et nettoyage des données de prix}
\usage{
prepare_price_data(df)
}
\arguments{
\item{df}{Un dataframe issu de import_matrix_prix}
}
\value{
Un dataframe nettoyé avec colonnes standardisées (Code_Produit, IP)
}
\description{
Préparation et nettoyage des données de prix
}
