% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01bis_projection_ratios.R
\name{trimestrialiser_cnt_complet}
\alias{trimestrialiser_cnt_complet}
\title{Trimestrialisation Complète (Historique + Projection)}
\usage{
trimestrialiser_cnt_complet(
  ratios_annuels,
  df_prix_niv3,
  derniere_annee_cna,
  annee_fin_proj
)
}
\arguments{
\item{ratios_annuels}{les ratios annuels deja calculé du TRE}

\item{df_prix_niv3}{les indices de prix au niveau le plus detaillé}

\item{derniere_annee_cna}{la dernière année des comptes definitifs}

\item{annee_fin_proj}{l'année de compilation}
}
\description{
Trimestrialisation Complète (Historique + Projection)
}
