% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/17_modele_equilibrage_ere.R
\name{charger_modele_equilibrage_ere}
\alias{charger_modele_equilibrage_ere}
\title{Charger le modèle d'équilibrage ERE sauvegardé}
\usage{
charger_modele_equilibrage_ere(
  path_rds = "C:/CnaBfaScn08/CntBfaV4/07P_Outils/OutilCntBfa/Modele_Equilibrage_ERE.rds",
  validate = TRUE
)
}
\arguments{
\item{path_rds}{Chemin vers le fichier \code{.rds}.}

\item{validate}{Si \code{TRUE} (défaut), vérifie la structure minimale.}
}
\value{
Liste de paramétrage du modèle d'équilibrage ERE.
}
\description{
Lit le fichier \code{.rds} de paramétrage préparé en amont et valide sa structure minimale.
}
