% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_utils.R
\name{clean_res}
\alias{clean_res}
\title{Consolidation des Résultats de benchmarking}
\usage{
clean_res(df, type_traitement)
}
\arguments{
\item{df}{la base de fichers benchmarkés}

\item{type_traitement}{est le type de traitement concerné}
}
\description{
Consolidation des Résultats de benchmarking
}
