% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_utils.R
\name{extract_price_branch}
\alias{extract_price_branch}
\title{Extrait la branche fine pour le Prix}
\usage{
extract_price_branch(code)
}
\arguments{
\item{code}{le code concaténé dont on veut extraire seulement le code branche
Ex: "Ind1_PRIVE*A01007" -> "A01007"
Ex: "A01007" -> "A01007" (Gère le cas où c'est déjà propre)}
}
\description{
Extrait la branche fine pour le Prix
}
