% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/13_ere_emplois.R
\name{lire_methodes_ere}
\alias{lire_methodes_ere}
\title{Lire et normaliser les methodes d estimation ERE depuis Methode_ERE.xlsx}
\usage{
lire_methodes_ere(path_methode_ere, map_feuille_cna)
}
\arguments{
\item{path_methode_ere}{Chemin vers le fichier Methode_ERE.xlsx.}

\item{map_feuille_cna}{Tibble de correspondance feuille <-> composante_cna
(colonnes : feuille, composante_cna).}
}
\value{
Tibble avec colonnes feuille, Code_Produit, Methode, composante_cna.
}
\description{
Lire et normaliser les methodes d estimation ERE depuis Methode_ERE.xlsx
}
